requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
chumpy==0.70
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
filelock==3.12.0
fonttools==4.39.3
freetype-py==2.4.0
frozenlist==1.3.3
fsspec==2023.5.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
idna==3.4
imageio==2.25.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lightning-utilities==0.8.0
lit==16.0.1
loguru==0.6.0
lpips==0.1.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.6.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.23.1
oauthlib==3.2.2
opencv-python==4.7.0.68
packaging==23.1
pandas==1.5.3
Pillow==9.4.0
plyfile==0.7.4
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyglet==2.0.9
PyOpenGL==3.1.0
pyparsing==3.0.9
pyrender==0.1.45
python-dateutil==2.8.2
pytorch-lightning==2.0.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
six==1.16.0
smplx==0.1.28
sympy==1.11.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==2.0.0
torchaudio==2.0.1
torchmetrics==0.11.4
torchvision==0.15.1
tqdm==4.64.1
trimesh==3.23.1
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yacs==0.1.8
yarl==1.9.2
